prevent FutureWarnings (#231)
    Category: bugfix
    JIRA issue: MIC-3488

pandas will show a FutureWarning when passing a list of length 1 to groupby. There is also a FutureWarning about using df.loc[] the way we did in updating the population.
Testing

Ran a small simulation all the way through and checked that groupby objects produced the same results using new and old methods. Checked that change to the population view column update produced the same behavior as the old code.